fixed compilation issues and added header file for tasks 1&2
