+++ software/frequency_relay_bsp/settings.bsp
@@ -2,9 +2,9 @@
 <sch:Settings xmlns:sch="http://www.altera.com/embeddedsw/bsp/schema">
         <BspType>hal</BspType>
         <BspVersion>default</BspVersion>
-        <BspGeneratedTimeStamp>18/04/2026 11:03:12 PM</BspGeneratedTimeStamp>
-        <BspGeneratedUnixTimeStamp>1776510192217</BspGeneratedUnixTimeStamp>
-        <BspGeneratedLocation>C:\Users\tpan328\Downloads\frequency-relay\software\frequency_relay_bsp</BspGeneratedLocation>
+        <BspGeneratedTimeStamp>19/04/2026 9:10:15 PM</BspGeneratedTimeStamp>
+        <BspGeneratedUnixTimeStamp>1776589815981</BspGeneratedUnixTimeStamp>
+        <BspGeneratedLocation>C:\Users\Tania\Documents\repos\frequency-relay\software\frequency_relay_bsp</BspGeneratedLocation>
         <BspSettingsFile>settings.bsp</BspSettingsFile>
         <SopcDesignFile>..\..\nios2.sopcinfo</SopcDesignFile>
         <JdiFile>default</JdiFile>
