+++ software/frequency_relay_bsp/summary.html
@@ -22,13 +22,13 @@
 <td width="20%" bgcolor="#77BBFF">BSP Version:</td><td>default</td>
 </tr>
 <tr mode="wrap">
-<td width="20%" bgcolor="#77BBFF">BSP Generated On:</td><td>18/04/2026 11:03:12 PM</td>
+<td width="20%" bgcolor="#77BBFF">BSP Generated On:</td><td>19/04/2026 9:10:15 PM</td>
 </tr>
 <tr mode="wrap">
-<td width="20%" bgcolor="#77BBFF">BSP Generated Timestamp:</td><td>1776510192217</td>
+<td width="20%" bgcolor="#77BBFF">BSP Generated Timestamp:</td><td>1776589815981</td>
 </tr>
 <tr mode="wrap">
-<td width="20%" bgcolor="#77BBFF">BSP Generated Location:</td><td>C:\Users\tpan328\Downloads\frequency-relay\software\frequency_relay_bsp</td>
+<td width="20%" bgcolor="#77BBFF">BSP Generated Location:</td><td>C:\Users\Tania\Documents\repos\frequency-relay\software\frequency_relay_bsp</td>
 </tr>
 </table>
 <br>
